MySQL/requirements.txt
mysql-replication==0.23
PyMySQL==0.9.3
